+++ ChangeLog.md
@@ -9,6 +9,7 @@
 * Blessure sanglante (voie du pacte vampirique)
 
 ### Autres améliorations
+* Prédicats pour les armes.
 * Tentative pour un meilleur suivi de la lumière en cas de changement de carte.
 * Meilleure prise en compte des créatures mortes qui régénèrent
 * Nouvelle syntaxe pour les prédicats à valeur complexe.